Bad request handling and adding http response for rest api's

+++ src/main/java/com/example/journalApp/Controller/JournalEntryControllerv2.java
@@ -4,13 +4,12 @@ import com.example.journalApp.Entity.JournalEntry;
 import com.example.journalApp.service.JournalEntryService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @RestController
 @RequestMapping("/journal/v2")
@@ -28,36 +27,55 @@ public class JournalEntryControllerv2 {
     }
 
     @GetMapping("/id/{id}")
-    public JournalEntry getJournalEntryById(@PathVariable ObjectId id)
+    public ResponseEntity<JournalEntry> getJournalEntryById(@PathVariable ObjectId id)
     {
-        return journalEntryService.getEntryById(id).orElse(null);
+        Optional<JournalEntry> entry = journalEntryService.getEntryById(id);
+        if(entry.isPresent())
+        {
+            return new ResponseEntity<>(entry.get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping
-    public JournalEntry CreateEntry(@RequestBody JournalEntry myEntry)
+    public ResponseEntity<JournalEntry> CreateEntry(@RequestBody JournalEntry myEntry)
     {
-        myEntry.setDate(LocalDateTime.now());
-        journalEntryService.saveEntry(myEntry);
-        return myEntry;
+        if(Objects.requireNonNull(myEntry).getTitle() != null || Objects.requireNonNull(myEntry).getContent() != null){
+            myEntry.setDate(LocalDateTime.now());
+            journalEntryService.saveEntry(myEntry);
+            return new ResponseEntity<>(myEntry,HttpStatus.CREATED);
+        }
+        else
+        {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @DeleteMapping("/id/{id}")
-    public void DeleteEntryById(@PathVariable ObjectId id)
+    public ResponseEntity<?> DeleteEntryById(@PathVariable ObjectId id)
     {
-        journalEntryService.DeleteEntryById(id);
+        JournalEntry myEntry = journalEntryService.getEntryById(id).orElse(null);
+        if(myEntry != null) {
+            journalEntryService.DeleteEntryById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PutMapping("/id/{id}")
-    public JournalEntry updateEntryById(@PathVariable ObjectId id, @RequestBody JournalEntry NewEntry)
+    public ResponseEntity<?> updateEntryById(@PathVariable ObjectId id, @RequestBody JournalEntry NewEntry)
     {
         JournalEntry myEntry = journalEntryService.getEntryById(id).orElse(null);
         if(myEntry != null)
         {
             myEntry.setTitle(NewEntry.getTitle() != null && !NewEntry.getTitle().isEmpty() ? NewEntry.getTitle() : myEntry.getTitle());
             myEntry.setContent(NewEntry.getContent() != null && !NewEntry.getContent().isEmpty() ? NewEntry.getContent() : myEntry.getContent());
+            journalEntryService.saveEntry(myEntry);
+            return new ResponseEntity<>(myEntry,HttpStatus.OK);
         }
-        journalEntryService.saveEntry(myEntry);
-        return myEntry;
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
 }
